Review KeepDistance
After first test run of this code on the robot, `KeepDistance` has changed to avoid the `not updating frequently` error and it is confirmed actuating on motors. Unfortunatelly It's still missing the mark. However I've added a new `!isMoving` condition to `KeepMoving` finishing condition. So the Is will not just set motor power to 0 and leave the motors free. it must confirm the drive train is still. This need testing I don't know how sensitive the accelerometers are.

I've also added a bunch of SmartDashboard Readings for debug.